backend/requirements.txt
# Insurance Chatbot Backend Dependencies
# Install with: pip install -r requirements.txt

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6  # For file uploads

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# RAG and embeddings
sentence-transformers==2.2.2  # Local embeddings (no API key needed)
faiss-cpu==1.7.4  # Vector store (use faiss-gpu for GPU support)
langchain==0.0.350  # LLM framework (optional, for future LLM integration)
numpy==1.24.3

# OpenAI for GPT-4 LLM
openai==1.3.0  # For GPT-4 LLM (required for production)

# Optional: For real STT/TTS providers (uncomment as needed)
# boto3==1.29.0  # For Amazon Polly TTS
# boto3==1.29.0  # For Amazon Polly TTS
# google-cloud-speech==2.21.0  # For Google Speech-to-Text
# google-cloud-texttospeech==2.14.0  # For Google TTS

# Development tools (optional)
# python-dotenv==1.0.0  # For .env file support

